+++ app/views/confirmation_mailer/alert.html.haml
