+++ app/views/confirmation_mailer/alert.text.erb
