--- config/initializers/email_confirmable.rb
@@ -1,1 +1,0 @@
-require 'email_confirmable/email_confirmable'
