--- lib/email_confirmable/app/controllers/email_confirmable/confirm_controller.rb
@@ -1,15 +1,0 @@
-class EmailConfirmable::ConfirmController < ApplicationController
-  # TODO: This is no longer used so maybe we should remove it
-  def confirmed
-    object = params[:resource].classify.constantize.find_by_confirm_id(params[:id])
-    if object
-      object.confirm!
-      singular = params[:resource].singularize
-      # This seems a bit long-winded. Is there a better way?
-      instance_variable_set(("@" + singular).to_sym, object)
-      render singular
-    else
-      raise ActiveRecord::RecordNotFound
-    end
-  end
-end
