--- lib/email_confirmable/email_confirmable.rb
@@ -1,8 +1,0 @@
-require File.join(File.dirname(__FILE__), 'email_confirmable', 'acts_as_email_confirmable')
-
-%w{ models controllers }.each do |dir|
-  path = File.join(File.dirname(__FILE__), 'app', dir)
-  $LOAD_PATH << path
-  ActiveSupport::Dependencies.autoload_paths << path
-  ActiveSupport::Dependencies.autoload_once_paths.delete(path)
-end
